moved some of the routes around
