+++ backend/money_maker/app.py
@@ -5,6 +5,7 @@ from money_maker.auth.routes import auth_bp
 from money_maker.extensions import (bcrypt, cache, celery, cors, db,
                                     jwt_manager, marshmallow)
 from money_maker.home.routes import home_bp
+from money_maker.models.user import User
 from money_maker.news.routes import news_stories_bp
 from money_maker.portfolio.routes import portfolio_bp
 from money_maker.quote.routes import quote_bp
@@ -118,3 +119,20 @@ class HTTPMethodOverrideMiddleware(object):
         if method in self.bodyless_methods:
             environ['CONTENT_LENGTH'] = '0'
         return self.app(environ, start_response)
+
+
+# Register a callback function that takes whatever object is passed in as the
+# identity when creating JWTs and converts it to a JSON serializable format.
+@jwt_manager.user_identity_loader
+def user_identity_lookup(user):
+    return user
+
+
+# Register a callback function that loads a user from your database whenever
+# a protected route is accessed. This should return any python object on a
+# successful lookup, or None if the lookup failed for any reason (for example
+# if the user has been deleted from the database).
+@jwt_manager.user_lookup_loader
+def user_lookup_callback(_jwt_header, jwt_data):
+    identity = jwt_data["sub"]
+    return db.session.query(User).filter(User.user_id == identity).one_or_none()
