+++ backend/money_maker/auth/routes.py
@@ -116,19 +116,3 @@ def which_user() -> flask.Response:
         return make_response(jsonify({"error": "error finding user"}), 400)
     return users_schema.jsonify(user)
 
-
-# Register a callback function that takes whatever object is passed in as the
-# identity when creating JWTs and converts it to a JSON serializable format.
-@jwt_manager.user_identity_loader
-def user_identity_lookup(user):
-    return user
-
-
-# Register a callback function that loads a user from your database whenever
-# a protected route is accessed. This should return any python object on a
-# successful lookup, or None if the lookup failed for any reason (for example
-# if the user has been deleted from the database).
-@jwt_manager.user_lookup_loader
-def user_lookup_callback(_jwt_header, jwt_data):
-    identity = jwt_data["sub"]
-    return db.session.query(User).filter(User.user_id == identity).one_or_none()
